Make changes to render category filter inside header component
