+++ src/containers/BooksForm.js
@@ -32,7 +32,7 @@ const BooksForm = () => {
       <h2 className="add-book"> ADD NEW BOOK </h2>
       <form onSubmit={(e) => handleChange(e)}>
         <input className="book-input" type="text" id="title" name="title" placeholder="Book Title" onChange={(e) => setTitle(e.target.value)} />
-        <Select className="select-dropdown-2" options={option} onChange={(e) => setCategory(e.value)} />
+        <Select className="select-dropdown-2" placeholder="Categories" options={option} onChange={(e) => setCategory(e.value)} />
         <button className="form-submit btn btn btn-lg" type="submit">Submit</button>
       </form>
     </div>
