+++ src/containers/BooksList.js
@@ -1,26 +1,11 @@
-import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import Book from '../components/Book';
-import { changeFilter, removeBook } from '../actions';
-import CategoryFilter from '../components/Categoryfilter';
+import { removeBook } from '../actions';
 
 const BooksList = () => {
   const { books, filter } = useSelector((state) => state);
-  const [options, setOptions] = useState();
 
   const dispatch = useDispatch();
-
-  useEffect(() => {
-    const categories = ['All', ...books.map((book) => book.category)];
-    setOptions(categories.map((category) => ({
-      value: category,
-      label: category,
-    })));
-  }, [books]);
-
-  const handleFilterChange = (e) => {
-    dispatch(changeFilter(e.value));
-  };
 
   const handleRemoveBook = (book) => {
     dispatch(removeBook(book));
@@ -39,11 +24,8 @@ const BooksList = () => {
 
   return (
     <div className="booklist">
-      <CategoryFilter options={options} handleChange={handleFilterChange} />
       <div className="booklist-table">
-
         {displayFilteredBooks()}
-
       </div>
     </div>
   );
